Move ByColor out
